requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.1.0

# Security
cryptography==42.0.2

# Rate limiting
Flask-Limiter==3.5.1

# Additional security headers
flask-talisman==1.1.0

# Background tasks
APScheduler==3.10.4
